requirements.txt
certifi==2020.6.20
charset-normalizer==2.0.5
cycler==0.10.0
dataclasses==0.6
future==0.18.2
idna==3.2
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.3.2
numpy==1.18.5
opencv-python==4.4.0.44
pandas==1.1.3
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.26.0
six==1.15.0
torch==1.7.0
torchvision==0.8.1
typing-extensions==3.7.4.3
urllib3==1.26.6
wget==3.2
